Update admin.svg - too large icon.
Sorry. I can not speak English very well. I rely on Google Translate.
There was a symptom of an icon appearing large on the screen because you did not specify the size in admin.svg.
Fix it.
I respect you. I want to contribute somehow, so I challenge for the first time.